Deprecate split-gpg and split-ctap
For some reason neither work. split-gpg2 spams qrexec asks and then does
nothing about it and split-ctap just doesn't seem to do anything. I've
tried to troubleshoot both today and split-gpg in multiple other
ocations and there is no progress.

Since they are not even QoL changes I'll forget them for a while and do
things manually with my Yubikey.

Note: there might be a bug in qubesos's GUI config editor because it
neves saves the configs for ctap (allow specific qubes to use all keys
or to generate keys).

And the fact that the GUI config editor uses split-gpg and the documentation
tells you to use split-gpg2 is quite bad.

Hopefully this gets resolved in 4.3 (if not I'll make a forum post and a
github issue).